Fixed Headcrab Bug eating while being dead. Fixed name
There was a bug with the headcrabs eating while already being dead due to a missing check. This check has been added.

During testing, we also found that the headcrabs are sometimes using the wrong name for what they are eating, because that name was based on its current/last target.